refactor(frontend): move dashboard sidebar and mobile navigation component to their respective file
- add sticky positioning to sidebar
